kc-connectors/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kc-connectors"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "kc-core",
    "kc-sandbox",
    "kc-supervisor",
    "python-telegram-bot>=21.0",
    "google-api-python-client>=2.130",
    "google-auth-oauthlib>=1.2",
    "google-auth-httplib2>=0.2",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "respx>=0.21", "ruff>=0.4"]

[tool.hatch.build.targets.wheel]
packages = ["src/kc_connectors"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v"
